Move all functions related to getting user info into one file
