--- server/src/actions/getSingleConnectedUser.ts
@@ -1,11 +1,0 @@
-import { connectedUsersCollection } from '../utilities/firebaseInit'
-
-export const getConnectedUserDisplayName = async (socketID: string) => {
-    try {
-        return connectedUsersCollection.doc(socketID).displayName; 
-    } catch (error) {
-        console.error(error);
-        return null;
-    }
-}
-
